#23 Refactor codes
