+++ META-INF/MANIFEST.MF
@@ -3,7 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: db2-editor
 Bundle-SymbolicName: plugin.db2;singleton:=true
 Bundle-Vendor: https://github.com/hamneko
-Bundle-Version: 0.3.0
+Bundle-Version: 0.3.1
 Require-Bundle: org.eclipse.cdt.core,
  org.eclipse.core.runtime,
  plugin.common,
