+++ pom.xml
@@ -6,7 +6,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>plugin.db2</groupId>
   <artifactId>plugin.db2</artifactId>
-  <version>0.3.0</version>
+  <version>0.3.1</version>
   <packaging>eclipse-plugin</packaging>
   <properties>
     <tycho.version>4.0.12</tycho.version>
